+++ test/test_helper.exs
@@ -1,11 +1,16 @@
 defmodule TestHelper do
   @fixture_project_dir "#{Path.dirname(__ENV__.file)}/fixture_projects"
+  @component_type_conversions_source_dir "#{@fixture_project_dir}/component_type_conversions"
   @wasm_test_source_dir "#{@fixture_project_dir}/wasm_test"
   @wasm_link_test_source_dir "#{@fixture_project_dir}/wasm_link_test"
   @wasm_link_dep_test_source_dir "#{@fixture_project_dir}/wasm_link_dep_test"
   @wasm_link_import_test_source_dir "#{@fixture_project_dir}/wasm_link_import_test"
   @wasm_import_test_source_dir "#{@fixture_project_dir}/wasm_import_test"
   @wasi_test_source_dir "#{@fixture_project_dir}/wasi_test"
+
+  def component_type_conversions_file_path,
+    do:
+      "#{@component_type_conversions_source_dir}/target/wasm32-wasip1/debug/component_type_conversions.wasm"
 
   def wasm_test_file_path,
     do: "#{@wasm_test_source_dir}/target/wasm32-unknown-unknown/debug/wasmex_test.wasm"
@@ -58,6 +63,9 @@ defmodule TestHelper do
         ],
         cd: @wasm_link_dep_test_source_dir
       )
+
+    {"", 0} =
+      System.cmd("cargo", ["component", "build"], cd: @component_type_conversions_source_dir)
   end
 
   def wasm_module do
@@ -116,6 +124,16 @@ defmodule TestHelper do
   def default_imported_functions_env_stringified do
     default_imported_functions_env()
     |> Enum.map(fn {k, v} -> {to_string(k), v} end)
+    |> Enum.into(%{})
+  end
+
+  def component_type_conversions_import_map do
+    ~w(
+      string u8 u16 u32 u64 s8 s16 s32 s64 f32 f64 bool char variant enum option-u8
+      result-u8-string result-u8-none result-none-string result-none-none
+      list-u8 tuple-u8-string flags point record-complex
+    )
+    |> Enum.map(fn id -> {"import-id-#{id}", {:fn, & &1}} end)
     |> Enum.into(%{})
   end